refactor: reverse hud avatar element
