+++ src/pages/dashboard/DashboardLayout.tsx
@@ -1,4 +1,3 @@
-import { DashboardHeader } from "@/component/header/DashboardHeader";
 import { Sidebar } from "@/component/sidebar/Sidebar";
 import { Outlet } from "react-router-dom";
 
@@ -6,7 +5,6 @@ export const DashboardLayout = () => {
   return (
     <div className="min-h-screen bg-gradient-to-br from-slate-900 via-slate-800 to-slate-900">
       <Sidebar />
-      <DashboardHeader />
       <Outlet />
     </div>
   );
